requirements.txt
streamlit
pillow
PyPDF2